school_awareness_platform/requirements.txt
Django
djangorestframework
django-cors-headers
psycopg2-binary
celery
redis
firebase-admin
django-jazzmin
djangorestframework-simplejwt
django-i18nfield
pillow
python-decouple